restructure to use go modules properly
