+++ .gitignore
@@ -1,0 +1,1 @@
+4600proj2
