+++ builtins/bash.go
