+++ builtins/bash_test.go
@@ -2,36 +2,37 @@ package builtins_test
 
 import (
 	"errors"
+	"os"
 	"os/exec"
 	"testing"
 
-	"github.com/jh125486/CSCE4600/Project2/builtins"
+	"github.com/BrandonAlexanderLopez/4600proj2/builtins"
 )
 
-func TestTcsh(t *testing.T) {
+func TestBash(t *testing.T) {
 	tests := []struct {
 		name    string
-		command []string
+		command string
 		wantErr error
 	}{
 		{
 			name:    "success echo",
-			command: []string{"echo", "hello world"},
+			command: "echo hello world",
 		},
 		{
 			name:    "success ls",
-			command: []string{"ls"},
+			command: "ls",
 		},
 		{
 			name:    "failure",
-			command: []string{"non-existing-command"},
+			command: "non-existing-command",
 			wantErr: &exec.Error{},
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if err := builtins.Tcsh(tt.command...); !errors.Is(err, tt.wantErr) {
-				t.Errorf("Tcsh() error = %v, wantErr %v", err, tt.wantErr)
+			if err := builtins.Bash(os.Stdout, tt.command); !errors.Is(err, tt.wantErr) {
+				t.Errorf("Bash() error = %v, wantErr %v", err, tt.wantErr)
 			}
 		})
 	}
