+++ builtins/cd_test.go
@@ -2,9 +2,10 @@ package builtins_test
 
 import (
 	"errors"
-	"github.com/jh125486/CSCE4600/Project2/builtins"
 	"os"
 	"testing"
+
+	"github.com/BrandonAlexanderLopez/4600proj2/builtins"
 )
 
 func TestChangeDirectory(t *testing.T) {
