+++ builtins/csh_test.go
@@ -5,7 +5,7 @@ import (
 	"os/exec"
 	"testing"
 
-	"github.com/jh125486/CSCE4600/Project2/builtins"
+	"github.com/BrandonAlexanderLopez/4600proj2/builtins"
 )
 
 func TestCsh(t *testing.T) {
