+++ builtins/echo.go
