+++ builtins/echo_test.go
@@ -2,9 +2,10 @@ package builtins_test
 
 import (
 	"bytes"
+	"fmt"
 	"testing"
 
-	"github.com/jh125486/CSCE4600/Project2/builtins"
+	"github.com/BrandonAlexanderLopez/4600proj2/builtins"
 )
 
 func TestEcho(t *testing.T) {
