+++ builtins/tcsh_test.go
@@ -5,33 +5,33 @@ import (
 	"os/exec"
 	"testing"
 
-	"github.com/jh125486/CSCE4600/Project2/builtins"
+	"github.com/BrandonAlexanderLopez/4600proj2/builtins"
 )
 
-func TestKsh(t *testing.T) {
+func TestTcsh(t *testing.T) {
 	tests := []struct {
 		name    string
-		command string
+		command []string
 		wantErr error
 	}{
 		{
 			name:    "success echo",
-			command: "echo hello world",
+			command: []string{"echo", "hello world"},
 		},
 		{
 			name:    "success ls",
-			command: "ls",
+			command: []string{"ls"},
 		},
 		{
 			name:    "failure",
-			command: "non-existing-command",
+			command: []string{"non-existing-command"},
 			wantErr: &exec.Error{},
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if err := builtins.Ksh(tt.command); !errors.Is(err, tt.wantErr) {
-				t.Errorf("Ksh() error = %v, wantErr %v", err, tt.wantErr)
+			if err := builtins.Tcsh(tt.command...); !errors.Is(err, tt.wantErr) {
+				t.Errorf("Tcsh() error = %v, wantErr %v", err, tt.wantErr)
 			}
 		})
 	}
